Update PageController updating deleteMessage to only delete current user messages

+++ SimpleChatJava/src/br/com/chat/controller/PageController.java
@@ -150,6 +150,7 @@ public class PageController {
      * 
      * @param dataModel
      * @param selectedRow
+     * @param user
      */
     public void deleteMessage(TableModel dataModel, int selectedRow, User user){
         
@@ -158,26 +159,33 @@ public class PageController {
         }
         else{
             
-            
             int id = (int) dataModel.getValueAt(selectedRow, 0);
+            int userId = (int) dataModel.getValueAt(selectedRow, 1);
             
-            try{
-                
-                MessageDAO messageDao = new MessageDAO(dataSource);
-                Message message = new Message();
-                
-                message.setId(id);
-                
-                messageDao.delete(message);
-                
-                JOptionPane.showMessageDialog(null, "Message deleted sulcefully.");
-                
+            if(isCurrentUser(userId, user.getId())){
+            
+                try{
+
+                    MessageDAO messageDao = new MessageDAO(dataSource);
+                    Message message = new Message();
+
+                    message.setId(id);
+
+                    messageDao.delete(message);
+
+                    JOptionPane.showMessageDialog(null, "Message deleted sulcefully.");
+
+                }
+                catch(SQLException e){
+                    JOptionPane.showMessageDialog(null, "error in exclude data: "+e.getMessage());
+                }
+                catch(Exception e){
+                    JOptionPane.showMessageDialog(null, "error: "+e.getMessage());
+                }
+            
             }
-            catch(SQLException e){
-                JOptionPane.showMessageDialog(null, "error in exclude data: "+e.getMessage());
-            }
-            catch(Exception e){
-                JOptionPane.showMessageDialog(null, "error: "+e.getMessage());
+            else{
+                JOptionPane.showMessageDialog(null, "You can only delete messages you sent.");
             }
             
         }
